logging_play/requirements.txt
typer
toml
pyyaml
python-json-logger
structlog
colorama
rich
python-json-logger
loguru
